fix: ui for new exports
Added FF

risk: low
JIRA: F1-1117
